[IMP] add confirmation message when deleting a record in form view
bzr revid: xmo@openerp.com-20111110103650-pqss9bfejudyqkp2

+++ addons/web/static/src/js/view_form.js
@@ -387,10 +387,16 @@ openerp.web.FormView = openerp.web.View.extend( /** @lends openerp.web.FormView#
         var def = $.Deferred();
         $.when(this.has_been_loaded).then(function() {
             if (self.can_be_discarded() && self.datarecord.id) {
-                self.dataset.unlink([self.datarecord.id]).then(function() {
-                    self.on_pager_action('next');
-                    def.resolve();
-                });
+                if (confirm(_t("Do you really want to delete this record?"))) {
+                    self.dataset.unlink([self.datarecord.id]).then(function() {
+                        self.on_pager_action('next');
+                        def.resolve();
+                    });
+                } else {
+                    setTimeout(function () {
+                        def.reject();
+                    }, 0)
+                }
             }
         });
         return def.promise();
